api/registration/requirements.txt
appnope==0.1.3
arrow==1.2.1
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
cfgv==3.3.1
click==8.0.3
decorator==5.1.1
distlib==0.3.7
Django==4.2.4
django-nextjs==2.3.1
django-cors-headers==4.2.0
django-extensions==3.2.3
djangorestframework==3.14.0
executing==1.2.0
filelock==3.12.2
gitlint==0.17.0
gitlint-core==0.17.0
identify==2.5.26
ipdb==0.13.13
ipython==8.14.0
jedi==0.19.0
matplotlib-inline==0.1.6
nodeenv==1.8.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.9.1
pre-commit==2.19.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
sh==1.14.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
toml==0.10.2
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.7.1
virtualenv==20.24.2
wcwidth==0.2.6
